board swal

+++ client/src/views/Boards.vue
@@ -42,8 +42,25 @@ export default {
       this.newBoard = { title: "", description: "" };
     },
     deleteBoard(data) {
-      this.$store.dispatch("deleteBoard", data);
+      swal({
+        title: "Are you sure?",
+        text: "Once deleted, you will not be able to recover this list!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true
+      }).then(willDelete => {
+        if (willDelete) {
+          this.$store.dispatch("deleteBoard", data);
+
+          swal("Poof! Your list has been deleted!", {
+            icon: "success"
+          });
+        } else {
+          swal("Your list has not been deleted!");
+        }
+      });
     },
+
     logout() {
       this.$store.dispatch("logout");
     }
